Added LOTS of stuff, also a lot of (not-so-successful-bugfix-attempts) in comments; some stuff (the signal messages) is still used, but not needed and therefore ignored overhead.
Summary: I strongly believe that I fixed all of our
so-callexd-laser-bugs by adding a sequencenumber (and a sequencer that
uses it in "Ausgabe.java") to messages and "notify change"s that have
something to do with updating the gui (==that are called during
"SimBoard"'s "doRealPhase()" method.
Ausgabe now adds every of those messages to a queue (ordered by sequence
number) and acknowledges them immediatly; but actions triggerd by those
messages will be delayed until the sequencenumber indicates that it is
the right time to do so.
Also change the sequence at serverside so that information messages are
sent before "notify change"; that was also part to an earlier fix-attempt
I did not undo.